refactor: rename ui-library to shared-ui
