+++ packages/package-bundler/bin/addTailwindStylesImport.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-npm explore @kaizen/package-bundler -- node ./dist/presets/ui-library/bin/addTailwindStylesImport.js --packagePath="$PWD"
+npm explore @kaizen/package-bundler -- node ./dist/presets/shared-ui/bin/addTailwindStylesImport.js --packagePath="$PWD"
