+++ packages/package-bundler/bin/markStyleInjectAsPure.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-npm explore @kaizen/package-bundler -- node ./dist/presets/ui-library/bin/markStyleInjectAsPure.js --packagePath="$PWD"
+npm explore @kaizen/package-bundler -- node ./dist/presets/shared-ui/bin/markStyleInjectAsPure.js --packagePath="$PWD"
