+++ packages/package-bundler/src/presets/index.ts
@@ -1,2 +1,2 @@
 export * from "./default/index.js"
-export * from "./ui-library/index.js"
+export * from "./shared-ui/index.js"
