--- packages/package-bundler/src/presets/ui-library/index.ts
@@ -1,1 +1,0 @@
-export * from "./plugins-ui-library.js"
